requirements.txt
streamlit
pandas
google-generativeai
plotly